# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VCanCare API",
        "description": "VCanCare Backend APIs",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/nurse/bookings": {
            "get": {
                "tags": [
                    "Nurse Bookings"
                ],
                "summary": "List nurse bookings",
                "description": "Paginated list of bookings assigned to the nurse.",
                "operationId": "listNurseBookings",
                "responses": {
                    "200": {
                        "description": "Bookings retrieved"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/nurse/schedule": {
            "get": {
                "tags": [
                    "Nurse Bookings"
                ],
                "summary": "Get daily schedule",
                "description": "Get all sessions for a specific date.",
                "operationId": "getNurseSchedule",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Schedule retrieved"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/nurse/sessions/{session_id}/start": {
            "post": {
                "tags": [
                    "Nurse Bookings"
                ],
                "summary": "Start session with OTP",
                "description": "Verifies the 6-digit OTP from the patient to start the session.",
                "operationId": "startNurseSession",
                "parameters": [
                    {
                        "name": "session_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "otp"
                                ],
                                "properties": {
                                    "otp": {
                                        "type": "string",
                                        "example": "123456",
                                        "maxLength": 6,
                                        "minLength": 6
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Session started"
                    },
                    "422": {
                        "description": "Invalid OTP"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/nurse/sessions/{session_id}/end": {
            "post": {
                "tags": [
                    "Nurse Bookings"
                ],
                "summary": "End session",
                "description": "Completes a started session. If all sessions are done, booking completes and nurse gets paid.",
                "operationId": "endNurseSession",
                "parameters": [
                    {
                        "name": "session_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "notes": {
                                        "type": "string",
                                        "example": "Vitals stable, medicine administered."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Session completed"
                    },
                    "409": {
                        "description": "Session not started"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/nurse/bookings/{booking_id}/cancel": {
            "post": {
                "tags": [
                    "Nurse Bookings"
                ],
                "summary": "Cancel booking (nurse)",
                "description": "Nurse cancels booking. User gets full refund for remaining sessions. Nurse receives cancellation strike.",
                "operationId": "cancelNurseBooking",
                "parameters": [
                    {
                        "name": "booking_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "reason": {
                                        "type": "string",
                                        "example": "Personal emergency"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Booking cancelled"
                    },
                    "409": {
                        "description": "Cannot cancel"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/nurse/wallet": {
            "get": {
                "tags": [
                    "Nurse Wallet"
                ],
                "summary": "Get nurse wallet",
                "description": "Returns wallet balance and paginated transaction/payout history.",
                "operationId": "getNurseWallet",
                "responses": {
                    "200": {
                        "description": "Wallet details"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/nurse/wallet/withdraw": {
            "post": {
                "tags": [
                    "Nurse Wallet"
                ],
                "summary": "Request withdrawal to bank",
                "description": "Nurse requests withdrawal of earnings to bank account. Minimum ₹100.",
                "operationId": "requestNurseWithdrawal",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount",
                                    "account_name",
                                    "account_number",
                                    "ifsc"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 2550
                                    },
                                    "account_name": {
                                        "type": "string",
                                        "example": "Nurse Name"
                                    },
                                    "account_number": {
                                        "type": "string",
                                        "example": "1234567890"
                                    },
                                    "ifsc": {
                                        "type": "string",
                                        "example": "SBIN0001234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Withdrawal request created"
                    },
                    "422": {
                        "description": "Insufficient balance or invalid amount"
                    },
                    "409": {
                        "description": "Duplicate pending request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/nurse/wallet/withdrawals": {
            "get": {
                "tags": [
                    "Nurse Wallet"
                ],
                "summary": "List withdrawal history",
                "description": "Paginated list of nurse withdrawal requests.",
                "operationId": "listNurseWithdrawals",
                "responses": {
                    "200": {
                        "description": "Withdrawal history"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}